requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
better-profanity==0.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.5
docopt==0.6.2
filelock==3.8.0
Flask==2.0.3
Flask-WTF==0.15.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
password-validator==1.0
pipreqs==0.4.11
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySimpleGUI==4.60.3
pytest==7.1.2
python-dotenv==0.19.0
requests==2.28.1
SQLAlchemy==1.4.41
tomli==2.0.1
urllib3==1.26.12
virtualenv==20.16.3
Werkzeug==2.2.2
WTForms==3.0.1
yarg==0.1.9
